pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "mws"
version = "0.1.0"
description = "Agent-first CLI for Microsoft 365 (Graph API)"
readme = "README.md"
license = "MIT"
requires-python = ">=3.11"
dependencies = [
  "typer[all]>=0.12",
  "rich>=13",
  "httpx[http2]>=0.27",
  "msal>=1.28",
  "pyyaml>=6",
  "platformdirs>=4",
  "openapi-core>=0.19",
  "mcp>=1.0",
]

[project.scripts]
mws = "mws.cli:app"

[tool.hatch.build.targets.wheel]
packages = ["src/mws"]

[project.optional-dependencies]
dev = [
  "pytest>=8",
  "pytest-asyncio>=0.23",
  "respx>=0.21",
  "ruff>=0.4",
  "mypy>=1.10",
  "pytest-cov>=5",
]

[tool.pytest.ini_options]
testpaths = ["tests"]
asyncio_mode = "auto"

[tool.ruff]
target-version = "py311"
src = ["src", "tests"]
line-length = 100

[tool.ruff.lint]
select = ["E", "F", "I", "UP", "B", "SIM"]

[tool.mypy]
python_version = "3.11"
strict = true
packages = ["mws"]
mypy_path = "src"
